reworked cefglue serialization

+++ src/DSerfozo.RpcBindings.CefGlue/Common/Serialization/CefTypes.cs
@@ -1,6 +1,6 @@
 ﻿namespace DSerfozo.RpcBindings.CefGlue.Common.Serialization
 {
-    public enum CefTypes : int
+    public enum CefTypes : byte
     {
         Int64 = 1,
         Time = 2,
